Adds logic to format post created date

+++ app/src/main/java/com/dendron/redditclient/domain/model/Post.kt
@@ -6,5 +6,5 @@ data class Post(
     val author: String,
     val thumbnail: String?,
     val comments: Int,
-    val created: Int,
+    val created: Long,
 )
